Upload files to ''
A string S of lowercase English letters is given. We want to partition this string into as many parts as possible so that each letter appears in at most one part, and return a list of integers representing the size of these parts.
S will have length in range [1, 500].
S will consist of lowercase English letters ('a' to 'z') only.